+++ tests/test_inputs.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 import re
 
-from datetime import datetime
+from datetime import date, datetime
 
 import pytz
 
@@ -24,25 +24,66 @@ class InputTest(object):
             yield assert_raises, error, lambda: input(value)
 
 
-class Rfc822Test(InputTest):
+class Rfc822DatetimeTest(InputTest):
     def test_valid_values(self):
         values = [
-            ("Sat, 01 Jan 2011 00:00:00 -0000", datetime(2011, 1, 1, tzinfo=pytz.utc)),
-            ("Sat, 01 Jan 2011 23:59:59 -0000", datetime(2011, 1, 1, 23, 59, 59, tzinfo=pytz.utc)),
-            ("Sat, 01 Jan 2011 21:59:59 -0200", datetime(2011, 1, 1, 23, 59, 59, tzinfo=pytz.utc)),
+            ('Sat, 01 Jan 2011', datetime(2011, 1, 1, tzinfo=pytz.utc)),
+            ('Sat, 01 Jan 2011 00:00:00 -0000', datetime(2011, 1, 1, tzinfo=pytz.utc)),
+            ('Sat, 01 Jan 2011 23:59:59 -0000', datetime(2011, 1, 1, 23, 59, 59, tzinfo=pytz.utc)),
+            ('Sat, 01 Jan 2011 21:59:59 -0200', datetime(2011, 1, 1, 23, 59, 59, tzinfo=pytz.utc)),
         ]
         self.assert_values(inputs.datetime_from_rfc822, values)
 
+    def test_error(self):
+        with assert_raises(ValueError):
+            inputs.datetime_from_rfc822('Fake, 01 Jan 2011')
 
-class Iso8601Test(InputTest):
+
+class Iso8601DatetimeTest(InputTest):
     def test_valid_values(self):
         values = [
-            ("2011-01-01T00:00:00+00:00", datetime(2011, 1, 1, tzinfo=pytz.utc)),
-            ("2011-01-01T23:59:59+00:00", datetime(2011, 1, 1, 23, 59, 59, tzinfo=pytz.utc)),
-            ("2011-01-01T23:59:59.001000+00:00", datetime(2011, 1, 1, 23, 59, 59, 1000, tzinfo=pytz.utc)),
-            ("2011-01-01T23:59:59+02:00", datetime(2011, 1, 1, 21, 59, 59, tzinfo=pytz.utc)),
+            ('2011-01-01', datetime(2011, 1, 1, tzinfo=pytz.utc)),
+            ('2011-01-01T00:00:00+00:00', datetime(2011, 1, 1, tzinfo=pytz.utc)),
+            ('2011-01-01T23:59:59+00:00', datetime(2011, 1, 1, 23, 59, 59, tzinfo=pytz.utc)),
+            ('2011-01-01T23:59:59.001000+00:00', datetime(2011, 1, 1, 23, 59, 59, 1000, tzinfo=pytz.utc)),
+            ('2011-01-01T23:59:59+02:00', datetime(2011, 1, 1, 21, 59, 59, tzinfo=pytz.utc)),
         ]
         self.assert_values(inputs.datetime_from_iso8601, values)
+
+    def test_error(self):
+        with assert_raises(ValueError):
+            inputs.datetime_from_iso8601('2008-13-13')
+
+
+class Rfc822DateTest(InputTest):
+    def test_valid_values(self):
+        values = [
+            ('Sat, 01 Jan 2011', date(2011, 1, 1)),
+            ('Sat, 01 Jan 2011 00:00:00 -0000', date(2011, 1, 1)),
+            ('Sat, 01 Jan 2011 23:59:59 -0000', date(2011, 1, 1)),
+            ('Sat, 01 Jan 2011 21:59:59 -0200', date(2011, 1, 1)),
+        ]
+        self.assert_values(inputs.date_from_rfc822, values)
+
+    def test_error(self):
+        with assert_raises(ValueError):
+            inputs.date_from_rfc822('Fake, 01 Jan 2011')
+
+
+class Iso8601DateTest(InputTest):
+    def test_valid_values(self):
+        values = [
+            ('2011-01-01', date(2011, 1, 1)),
+            ('2011-01-01T00:00:00+00:00', date(2011, 1, 1)),
+            ('2011-01-01T23:59:59+00:00', date(2011, 1, 1)),
+            ('2011-01-01T23:59:59.001000+00:00', date(2011, 1, 1)),
+            ('2011-01-01T23:59:59+02:00', date(2011, 1, 1)),
+        ]
+        self.assert_values(inputs.date_from_iso8601, values)
+
+    def test_error(self):
+        with assert_raises(ValueError):
+            inputs.date_from_iso8601('2008-13-13')
 
 
 class UrlTest(InputTest):
@@ -227,56 +268,56 @@ class IsoIntervalTest(InputTest):
     def test_valid_values(self):
         values = [(
             # Full precision with explicit UTC.
-            "2013-01-01T12:30:00Z/P1Y2M3DT4H5M6S",
+            '2013-01-01T12:30:00Z/P1Y2M3DT4H5M6S',
             (
                 datetime(2013, 1, 1, 12, 30, 0, tzinfo=pytz.utc),
                 datetime(2014, 3, 5, 16, 35, 6, tzinfo=pytz.utc),
             ),
         ), (
             # Full precision with alternate UTC indication
-            "2013-01-01T12:30+00:00/P2D",
+            '2013-01-01T12:30+00:00/P2D',
             (
                 datetime(2013, 1, 1, 12, 30, 0, tzinfo=pytz.utc),
                 datetime(2013, 1, 3, 12, 30, 0, tzinfo=pytz.utc),
             ),
         ), (
             # Implicit UTC with time
-            "2013-01-01T15:00/P1M",
+            '2013-01-01T15:00/P1M',
             (
                 datetime(2013, 1, 1, 15, 0, 0, tzinfo=pytz.utc),
                 datetime(2013, 1, 31, 15, 0, 0, tzinfo=pytz.utc),
             ),
         ), (
             # TZ conversion
-            "2013-01-01T17:00-05:00/P2W",
+            '2013-01-01T17:00-05:00/P2W',
             (
                 datetime(2013, 1, 1, 22, 0, 0, tzinfo=pytz.utc),
                 datetime(2013, 1, 15, 22, 0, 0, tzinfo=pytz.utc),
             ),
         ), (
             # Date upgrade to midnight-midnight period
-            "2013-01-01/P3D",
+            '2013-01-01/P3D',
             (
                 datetime(2013, 1, 1, 0, 0, 0, tzinfo=pytz.utc),
                 datetime(2013, 1, 4, 0, 0, 0, 0, tzinfo=pytz.utc),
             ),
         ), (
             # Start/end with UTC
-            "2013-01-01T12:00:00Z/2013-02-01T12:00:00Z",
+            '2013-01-01T12:00:00Z/2013-02-01T12:00:00Z',
             (
                 datetime(2013, 1, 1, 12, 0, 0, tzinfo=pytz.utc),
                 datetime(2013, 2, 1, 12, 0, 0, tzinfo=pytz.utc),
             ),
         ), (
             # Start/end with time upgrade
-            "2013-01-01/2013-06-30",
+            '2013-01-01/2013-06-30',
             (
                 datetime(2013, 1, 1, tzinfo=pytz.utc),
                 datetime(2013, 6, 30, tzinfo=pytz.utc),
             ),
         ), (
             # Start/end with TZ conversion
-            "2013-02-17T12:00:00-07:00/2013-02-28T15:00:00-07:00",
+            '2013-02-17T12:00:00-07:00/2013-02-28T15:00:00-07:00',
             (
                 datetime(2013, 2, 17, 19, 0, 0, tzinfo=pytz.utc),
                 datetime(2013, 2, 28, 22, 0, 0, tzinfo=pytz.utc),
@@ -285,63 +326,63 @@ class IsoIntervalTest(InputTest):
         # Resolution expansion for single date(time)
         (
             # Second with UTC
-            "2013-01-01T12:30:45Z",
+            '2013-01-01T12:30:45Z',
             (
                 datetime(2013, 1, 1, 12, 30, 45, tzinfo=pytz.utc),
                 datetime(2013, 1, 1, 12, 30, 46, tzinfo=pytz.utc),
             ),
         ), (
             # Second with tz conversion
-            "2013-01-01T12:30:45+02:00",
+            '2013-01-01T12:30:45+02:00',
             (
                 datetime(2013, 1, 1, 10, 30, 45, tzinfo=pytz.utc),
                 datetime(2013, 1, 1, 10, 30, 46, tzinfo=pytz.utc),
             ),
         ), (
             # Second with implicit UTC
-            "2013-01-01T12:30:45",
+            '2013-01-01T12:30:45',
             (
                 datetime(2013, 1, 1, 12, 30, 45, tzinfo=pytz.utc),
                 datetime(2013, 1, 1, 12, 30, 46, tzinfo=pytz.utc),
             ),
         ), (
             # Minute with UTC
-            "2013-01-01T12:30+00:00",
+            '2013-01-01T12:30+00:00',
             (
                 datetime(2013, 1, 1, 12, 30, tzinfo=pytz.utc),
                 datetime(2013, 1, 1, 12, 31, tzinfo=pytz.utc),
             ),
         ), (
             # Minute with conversion
-            "2013-01-01T12:30+04:00",
+            '2013-01-01T12:30+04:00',
             (
                 datetime(2013, 1, 1, 8, 30, tzinfo=pytz.utc),
                 datetime(2013, 1, 1, 8, 31, tzinfo=pytz.utc),
             ),
         ), (
             # Minute with implicit UTC
-            "2013-01-01T12:30",
+            '2013-01-01T12:30',
             (
                 datetime(2013, 1, 1, 12, 30, tzinfo=pytz.utc),
                 datetime(2013, 1, 1, 12, 31, tzinfo=pytz.utc),
             ),
         ), (
             # Hour, explicit UTC
-            "2013-01-01T12Z",
+            '2013-01-01T12Z',
             (
                 datetime(2013, 1, 1, 12, tzinfo=pytz.utc),
                 datetime(2013, 1, 1, 13, tzinfo=pytz.utc),
             ),
         ), (
             # Hour with offset
-            "2013-01-01T12-07:00",
+            '2013-01-01T12-07:00',
             (
                 datetime(2013, 1, 1, 19, tzinfo=pytz.utc),
                 datetime(2013, 1, 1, 20, tzinfo=pytz.utc),
             ),
         ), (
             # Hour with implicit UTC
-            "2013-01-01T12",
+            '2013-01-01T12',
             (
                 datetime(2013, 1, 1, 12, tzinfo=pytz.utc),
                 datetime(2013, 1, 1, 13, tzinfo=pytz.utc),
@@ -349,7 +390,7 @@ class IsoIntervalTest(InputTest):
         ), (
             # Interval with trailing zero fractional seconds should
             # be accepted.
-            "2013-01-01T12:00:00.0/2013-01-01T12:30:00.000000",
+            '2013-01-01T12:00:00.0/2013-01-01T12:30:00.000000',
             (
                 datetime(2013, 1, 1, 12, tzinfo=pytz.utc),
                 datetime(2013, 1, 1, 12, 30, tzinfo=pytz.utc),